+++ lib/src/foundation/carousel/carousel.dart
@@ -1,8 +1,8 @@
 library;
 
-import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:tv_plus/src/foundation/foundation.dart';
+import 'package:flutter/widgets.dart';
+import 'package:tv_plus/src/foundation/dpad/dpad.dart';
 
 part 'tv_carousel_controller.dart';
 part 'tv_carousel_pager.dart';