+++ lib/src/foundation/foundation.dart
@@ -1,5 +1,6 @@
 library;
 
+export 'carousel/carousel.dart';
 export 'dpad/dpad.dart';
 export 'navigation/navigation.dart';
 export 'scroll/scroll.dart';