+++ lib/src/foundation/navigation/menu/menu.dart
@@ -8,4 +8,4 @@ part 'tv_navigation_menu_alignment.dart';
 part 'tv_navigation_menu_content.dart';
 part 'tv_navigation_menu_controller.dart';
 part 'tv_navigation_menu_item.dart';
-part 'tv_navigation_menu_selection_entry.dart';
+part 'tv_navigation_menu_entry.dart';