+++ lib/src/foundation/scroll/tv_grid_view.dart
@@ -15,7 +15,7 @@ final class TvGridView extends BoxScrollView {
     bool addRepaintBoundaries = true,
     bool addSemanticIndexes = true,
     super.cacheExtent,
-    List<ScrollGroupDpadFocus> children = const [],
+    List<Widget> children = const [],
     int? semanticChildCount,
     this.focusScopeNode,
     FocusTraversalPolicy? policy,
@@ -26,6 +26,8 @@ final class TvGridView extends BoxScrollView {
     this.onDown,
     this.onLeft,
     this.onRight,
+    this.onFocusChanged,
+    this.onFocusDisabledWhenWasFocused,
     super.dragStartBehavior,
     super.keyboardDismissBehavior,
     super.restorationId,
@@ -62,6 +64,8 @@ final class TvGridView extends BoxScrollView {
     this.onDown,
     this.onLeft,
     this.onRight,
+    this.onFocusChanged,
+    this.onFocusDisabledWhenWasFocused,
     bool addAutomaticKeepAlives = true,
     bool addRepaintBoundaries = true,
     bool addSemanticIndexes = true,
@@ -90,10 +94,12 @@ final class TvGridView extends BoxScrollView {
   final bool descendantsAreFocusable;
   final bool descendantsAreTraversable;
   final bool autofocus;
-  final ScrollGroupDpadEventCallback? onUp;
-  final ScrollGroupDpadEventCallback? onDown;
-  final ScrollGroupDpadEventCallback? onLeft;
-  final ScrollGroupDpadEventCallback? onRight;
+  final DpadScopeEventCallback? onUp;
+  final DpadScopeEventCallback? onDown;
+  final DpadScopeEventCallback? onLeft;
+  final DpadScopeEventCallback? onRight;
+  final void Function(FocusScopeNode, bool)? onFocusChanged;
+  final void Function(FocusScopeNode)? onFocusDisabledWhenWasFocused;
 
   @override
   Widget buildChildLayout(BuildContext context) {
@@ -104,11 +110,16 @@ final class TvGridView extends BoxScrollView {
   Widget build(BuildContext context) {
     return DpadFocusScope(
       focusScopeNode: focusScopeNode,
+      policy: policy,
+      descendantsAreFocusable: descendantsAreFocusable,
+      descendantsAreTraversable: descendantsAreTraversable,
       autofocus: autofocus,
       onUp: onUp,
       onDown: onDown,
       onLeft: onLeft,
       onRight: onRight,
+      onFocusChanged: onFocusChanged,
+      onFocusDisabledWhenWasFocused: onFocusDisabledWhenWasFocused,
       builder: (_) => super.build(context),
     );
   }