+++ lib/src/material/material.dart
@@ -1,2 +1,1 @@
-export 'carousel/carousel.dart';
 export 'navigation/navigation.dart';